fixed uploading book and book cover. Creating books successful.

+++ booking-social/src/main/java/com/ianclark226/booking_social/feedback/Feedback.java
@@ -25,7 +25,7 @@ import java.time.LocalDateTime;
 public class Feedback extends BaseEntity {
 
     @Column
-    private Double note;
+    private Integer note;
     private String comment;
     @ManyToOne
     @JoinColumn(name = "book_id")
